Move the kubernetes discovery from conf to modules/discovery
